Codigo/requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
python-dotenv==0.21.0
requests==2.27.1
selenium==4.4.3
webdriver_manager==3.5.4
